sort of working

+++ src/map.rs
@@ -1,3 +1,5 @@
+use std::isize;
+
 use crate::{Color, Drawable, FrameBuffer};
 
 #[derive(Debug)]
@@ -21,8 +23,8 @@ impl Ray {
     }
 
     pub fn length(&self) -> f32 {
-        let d_x = (self.start.x - self.end.x) as f32;
-        let d_y = (self.start.y - self.end.y) as f32;
+        let d_x = (self.start.x as isize - self.end.x as isize) as f32;
+        let d_y = (self.start.y as isize - self.end.y as isize) as f32;
 
         (d_x.powf(2.) + d_y.powf(2.)).sqrt()
     }
